Moved flying hit objects to separate files
