+++ osu.Game.Rulesets.Taiko/Objects/Drawables/DrawableCentreHit.cs
@@ -2,10 +2,7 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Allocation;
-using osu.Game.Beatmaps;
-using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Graphics;
-using osu.Game.Rulesets.Scoring;
 using osu.Game.Rulesets.Taiko.Objects.Drawables.Pieces;
 
 namespace osu.Game.Rulesets.Taiko.Objects.Drawables
@@ -24,20 +21,6 @@ namespace osu.Game.Rulesets.Taiko.Objects.Drawables
         private void load(OsuColour colours)
         {
             MainPiece.AccentColour = colours.PinkDarker;
-        }
-    }
-
-    public class DrawableFlyingCentreHit : DrawableCentreHit
-    {
-        protected override void CheckForResult(bool userTriggered, double timeOffset)
-        {
-            ApplyResult(r => r.Type = HitResult.Good);
-        }
-
-        public DrawableFlyingCentreHit(double time, bool isStrong = false)
-            : base(new IgnoreHit { StartTime = time, IsStrong = isStrong })
-        {
-            HitObject.ApplyDefaults(new ControlPointInfo(), new BeatmapDifficulty());
         }
     }
 }
